backend/requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
asgi-lifespan==1.0.1
asyncio==3.4.3
asyncpg==0.30.0
attrs==25.3.0
certifi==2025.4.26
click==8.1.8
colorama==0.4.6
coverage==7.8.2
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
factory_boy==3.3.3
Faker==37.3.0
fastapi==0.115.12
fastapi-cli==0.0.7
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
psycopg2-binary==2.9.10
pyasn1==0.4.8
pydantic==2.11.2
pydantic-settings==2.9.1
pydantic_core==2.33.1
Pygments==2.19.1
pytest==8.3.5
pytest-asyncio==0.18.3
pytest-cov==3.0.0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.2
rich==14.0.0
rich-toolkit==0.14.6
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.1
tomli==2.2.1
typer==0.15.4
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0.1
